# docs/swagger.yaml
definitions:
...
  dto.PlanDTO:
    properties:
      duration:
        type: integer
      max_amount:
        type: number
      min_amount:
        type: number
      name:
        type: string
      roi:
        type: number
    required:
    - duration
    - max_amount
    - min_amount
    - name
    - roi
    type: object
...
  response.ApiResponse:
    properties:
      data: {}
      error: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is my Go Inv API.
  title: My Inv API
  version: "1.0"
paths:
  /admin/dashboard:
    get:
      description: Returns dashboard data for admin dashboard
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Admin dashboard data
      tags:
      - Admin Dashboard
...
  /admin/plan/{id}:
    delete:
      consumes:
      - application/json
      description: Delete investment plan
      parameters:
      - description: Plan ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Delete plan
      tags:
      - Admin Action
  /admin/plans:
    post:
      consumes:
      - application/json
      description: Create new inv plan
      parameters:
      - description: plan data
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.PlanDTO'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Create new plan
      tags:
      - Admin Action
...
  /user/dashboard:
    get:
      description: Returns dashboard data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User dashboard page
      tags:
      - User Dashboard
  /user/deposits:
    get:
      description: Returns deposit page data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User deposit page data
      tags:
      - User Dashboard
    post:
      consumes:
      - multipart/form-data
      description: Create a new deposit for a user
      parameters:
      - description: Deposit Amount
        in: formData
        name: amount
        required: true
        type: number
      - description: Deposit Proof Image
        in: formData
        name: deposit_proof
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: New deposit
      tags:
      - User Action
...
  /user/kyc/upload:
    post:
      consumes:
      - multipart/form-data
      description: Upload user KYC ID front and back images
      parameters:
      - description: ID Front Image
        in: formData
        name: id_front
        required: true
        type: file
      - description: ID Back Image
        in: formData
        name: id_back
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: Upload KYC documents
      tags:
      - User Action
  /user/packages:
    get:
      description: Returns investment page data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User investment page data
      tags:
      - User Dashboard
...
  /user/transactions:
    get:
      description: Returns transaction page data for the logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ApiResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ApiResponse'
      security:
      - BearerAuth: []
      summary: User transaction page data
      tags:
      - User Dashboard
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"